ORM/Persistenz-Stack (aber funktionierend) hinzugefügt.

+++ src/main/java/com/example/BackendPlanti/DB.java
@@ -18,7 +18,7 @@ public class DB {
 
 
         session.beginTransaction();
-        Pflanze student = session.load(Pflanze.class, input.longValue());
+        Pflanze student = session.load(Pflanze.class, input);
         session.flush();
         return student;
     }
